+++ sample-apps/react/audio-rooms/src/pages/Login.tsx
@@ -1,5 +1,6 @@
 import icon from '../assets/icon.png';
 import { useNavigate } from 'react-router-dom';
+import { JSX } from 'react';
 import { useUserContext } from '../contexts';
 import users from '../data/users';